Add files via upload
The summary plot shows global feature importance. The sina plots show the distribution of feature contributions to the model output using SHAP values of each feature for every observation. Each dot is an observation

-SHAP Values: The x-axis represents the SHAP values, which measure the impact of each feature on the model's output. A SHAP value can be positive or negative, indicating whether the feature pushes the model's prediction higher or lower, respectively.

- Impact Direction: Features that are shown on the right side of the plot have a positive impact on the model's prediction, increasing the likelihood of a certain outcome. Features on the left decrease the likelihood.

this graph is a powerful tool to understand which features are most important in a model's predictions and how the values of those features change the predictions.